First commit
The multiboard feature is here to connect two marlin board via SPI in a master/slave configuration.

For now, i'm only working on implementing the PWM Fan control of the slave board from the master.

In this commit, i added some lines to allow for virtual fans to exist on the master, and to get the PWM value of those virtuals fans into my new class named "multiboard"